Add variable assignments
Variables can be declared and shadowing is supported, but modifying
existing variables via assignments was not available.

This meant that modifying a variable from a nested block was not
possible:

	{{$v := "init"}}
	{{if true}}
		{{$v := "changed"}}
	{{end}}
	v: {{$v}} {{/* "init" */}}

Introduce the "=" assignment token, such that one can now do:

	{{$v := "init"}}
	{{if true}}
		{{$v = "changed"}}
	{{end}}
	v: {{$v}} {{/* "changed" */}}

IMPORTANT: since the "=" variable declaration token has been changed to ":=", there is no backwards compatibility with version 1.0.